+++ src/pubify_mpl/__init__.py
@@ -1,18 +1,25 @@
-from .export import ResolvedStyle, save_fig
-from .layout import DEFAULT_TEMPLATE, use_template
+from .export import (
+    ResolvedStyle,
+    auto_rasterize_figure,
+    figure_renderer,
+    figure_tight_bbox,
+    prepare_figure,
+    save_fig,
+)
 from .rc import pubify_rc_context
-from .resources import install_pubify_package, prepare, write_tex_template
 from .adjust import remove_outside_padding
+from .style import DEFAULT_STYLE, normalized_style
 
 
 __all__ = [
-    "DEFAULT_TEMPLATE",
-    "install_pubify_package",
-    "prepare",
+    "DEFAULT_STYLE",
+    "auto_rasterize_figure",
+    "figure_renderer",
+    "figure_tight_bbox",
+    "normalized_style",
+    "prepare_figure",
     "pubify_rc_context",
     "remove_outside_padding",
     "ResolvedStyle",
     "save_fig",
-    "use_template",
-    "write_tex_template",
 ]